Rename default codegen to codegen_cranelift
